requirements.txt
pandas>=2.0.0
google-generativeai>=0.3.0
speechrecognition>=3.10.0
pyaudio>=0.2.11
python-dotenv>=1.0.0
flask>=2.3.0
pyinstaller>=5.13.0
openai>=1.0.0